+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonTest/kotlin/net/folivo/trixnity/clientserverapi/client/AppserviceApiClientTest.kt
@@ -16,6 +16,7 @@ class AppserviceApiClientTest {
     fun shouldPing() = runTest {
         val matrixRestClient = MatrixClientServerApiClientImpl(
             baseUrl = Url("https://matrix.host"),
+            authProvider = MatrixAuthProvider.classicInMemory("token"),
             httpClientEngine = scopedMockEngine {
                 addHandler { request ->
                     assertEquals("/_matrix/client/v1/appservice/appId/ping", request.url.fullPath)