+++ trixnity-clientserverapi/trixnity-clientserverapi-client/src/commonTest/kotlin/net/folivo/trixnity/clientserverapi/client/AuthApiClientTest.kt
@@ -16,7 +16,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-class AuthenticationApiClientTest {
+class AuthApiClientTest {
 
     @Test
     fun shouldGetWhoami() = runTest {