+++ trixnity-clientserverapi/trixnity-clientserverapi-server/src/commonTest/kotlin/net/folivo/trixnity/clientserverapi/server/MatrixAccessTokenAuthTest.kt
@@ -15,10 +15,11 @@ import io.ktor.server.routing.*
 import io.ktor.server.testing.*
 import kotlinx.serialization.Serializable
 import net.folivo.trixnity.api.server.matrixEndpoint
+import net.folivo.trixnity.core.Auth
+import net.folivo.trixnity.core.AuthRequired
 import net.folivo.trixnity.core.HttpMethod
 import net.folivo.trixnity.core.HttpMethodType.GET
 import net.folivo.trixnity.core.MatrixEndpoint
-import net.folivo.trixnity.core.WithoutAuth
 import net.folivo.trixnity.core.serialization.createDefaultEventContentSerializerMappings
 import net.folivo.trixnity.core.serialization.createMatrixEventJson
 import kotlin.test.Test
@@ -133,13 +134,13 @@ class MatrixAccessTokenAuthTest {
     @Serializable
     @Resource("/get")
     @HttpMethod(GET)
-    @WithoutAuth
+    @Auth(AuthRequired.NO)
     object GetResourceWithoutAuth : MatrixEndpoint<Unit, Unit>
 
     @Serializable
     @Resource("/get")
     @HttpMethod(GET)
-    @WithoutAuth(true)
+    @Auth(AuthRequired.OPTIONAL)
     object GetResourceWithOptionalAuth : MatrixEndpoint<Unit, Unit>
 
     @Test